File moved ```StudentProjectsDetailService```` FROM namespace App\Service\Resume; TO namespace App\Service\Student;
